requirements.txt
# Core MCP dependencies
mcp>=1.0.0

# LAMMPS Python interface
lammps>=2023.0.0

# Data processing and scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0

# Molecular chemistry and force field tools
rdkit>=2023.3.1
openbabel>=3.1.1
MDAnalysis>=2.4.0
pytraj>=2.0.5

# OpenFF Toolkit for force field assignment
openff-toolkit>=2.2.0
openff-interchange>=0.3.0
openff-units>=0.2.0
openff-models>=0.1.2
openff-nagl>=0.3.0

# Data validation and serialization
pydantic>=2.0.0

# File handling and configuration
pyyaml>=6.0
pathlib2>=2.3.0

# Async support
asyncio-mqtt>=0.11.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: GPU support
# cupy>=11.0.0  # Uncomment if CUDA support is needed 